requirements.txt
numpy
scikit-learn
matplotlib
sys
opencv-python
matplotlib
time
scipy.io